requirements.txt
# Ensure pip is up-to-date in Render build
pip>=25.3

# Core Django packages
Django>=5.0,<6.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.1

# WSGI server for production
gunicorn>=21.2.0

# Environment variable handling
python-dotenv>=1.0.1
python-decouple>=3.8

# Database URL parsing
dj-database-url>=2.2.0

# Static file handling
whitenoise>=6.6.0

# HTTP requests
requests>=2.31.0

# Payment SDK
paypalrestsdk>=1.13.1

# PostgreSQL driver (fully precompiled)
psycopg2-binary==2.9.9
